TimeScaleManager, bugfixes, organization
TimeScaleManager now exists, and is the only thing that should be directly touching the timescale of the game.

Further organized scripts into folders.

Fixed a bug where during the charge attack, the player colliding with anything (including the attacked enemy) could result in them being positioned incorrectly at the end of the dash.