Slight heirarchy organization improvement, spawn limits increase over time and with score
There is now a Difficulty Manager which creates a "difficulty" value to determine how much the survival mode should scale over time/score. The Spawn Manager now uses this value to increase spawn limits of enemy types.

There is now an empty "Enemies" object which should be in the heirarchy of any scene. This is for organizing the heirarchy, and any spawned enemies should exist within it.