Slightly less stupid difficulty assignment
Instead of using different scenes for different difficulties, we now assign difficulty values when we select a difficulty and pass them into the same scene, like sane people.

Created DifficultyValues to hold values for difficulty level.

Adjusted DifficultyManager, EnemyAttackTokenPool, and SpawnManager to use these values.

Extended GameObject with DoDestroyOnLoad() because why doesn't that already exist, and also we need it since Difficulty Values isn't actually a singleton in the menus, but we do want it to be destroyed once we leave the level, but not when we leave the menu scene.

Created Pause canvas for later use.