Events System and pausing updates
There's now an event system! Technically you can pass an arbitrary amount of data through the events system, but it's really just a string for now. Later we could add support for passing JSON through the string that can be parsed into scriptable game objects and stuff, if we need it.

The pause stuff also is now decoupled from the PlayerMover and uses the new EventManager instead, and the pause menu has a Resume button.