adds rudimentary ability for spiders to climb, fixes error with pickups
Spiders will climb once they hit a surface whose normal is not Vector3.up. Spiders climb until the reach the top of the surface, even if it is not in the direction of their path - they just need to collide with it. Surface detection is also currently buggy, causing spiders to frequently "jump"